fix(capture): the label takes the core's own spelling, and the wiring is asserted
Review findings on the capture-source directory, applied.

**The published label went through titleCasedID and skipped the compiled-in
names.** A unit declaring ingress from `whatsapp` published "Whatsapp" in the
same document where `data` carries the core's "WhatsApp" — two spellings of one
transport in one list, which is the defect this seam exists to remove. It goes
through providerLabel now, so a compiled-in spelling wins wherever there is one
and a unit-only key still falls through to the derived name.

A system key matching a core transport is still not refused, unlike a unit's
CHANNEL declaration, and the comment now says why: a channel claims "I transmit
on this", where a unit taking a core transport would re-point the core's row at
the unit's own credential; an ingress source claims "I bring records in from
this", and a unit ingesting Telegram exports naming it `telegram` is telling the
truth. Refusing it would break the honest case to stop a reviewed first-party
unit from lying about what it captures.

**titleCasedID had started collapsing empty segments.** Widening it from `_` to
both separators swapped Split for FieldsFunc, which drops empty fields — so
`deal__room` became "Deal Room" where migration 0247's
`initcap(replace(provider,'_',' '))` seeds "Deal  Room". Both are legal under the
column's CHECK, so a fresh install and a booted one would have disagreed about
that provider's name, which is the one thing the seeded-label fitness test
promises cannot happen. `-` is folded onto `_` instead, keeping one separator and
the empty-segment behaviour.

**Nothing asserted that the handler serves the field.** Every test shaped a slice
handed in by the caller, leaving `publishedCaptureSources(ComposedExtensions())`
— the line the whole seam rides on — covered by nothing: drop it and every gate
stayed green. There is now a test through the real composed set and the real
handler, reading the response bytes rather than the struct, because a shaping
that returns nil and a handler that drops the key are the same bug from a
client's side. Verified failing with the line removed.

The pattern check is derived from the grammars boot admits rather than sampled
from three hand-written units: a conforming client may refuse a whole response
over one bad entry, so the question is what CAN be admitted, not what is composed
today.

**The contract claimed a completeness it does not have.** `capture_sources` is
not every non-transport connector a trace can name — a core connector's own
source system (`gmail`, `imap`, `graph`, `gcal`) is in neither array, and a
client that treats a directory miss as an anomaly would report the most common
trace row in the product. Said plainly now, and `label`'s description keeps the
client-facing half of its rationale and drops the server's authorization
argument, which already lives in three places in Go.

**The endpoint's authorization argument only defended half of itself.** It argued
label provenance; what this change widened is the SET, since `capture_sources`
publishes which units are composed and what each captures from. Both halves are
stated now, on the ground LoadChannelProviderDirectory already records — what is
gated is the act, not the name — including the one thing genuinely new, that a
capture-only unit becomes visible to a seat holding no grant on it.

Tests move to the files their subjects live in, and three comments that narrated
the change now state the invariant instead.
