+++ backend/internal/compose/extingress.go
@@ -358,16 +358,16 @@ func (r *callRuntime) sourceSystem(system string) string {
 	return extSourceSystem(r.unit, system)
 }
 
-// extSourceSystem spells the provenance a unit's landed record carries, and it
-// is spelled HERE ONLY.
+// extSourceSystem spells the provenance a unit's landed record carries. One
+// spelling in product code; the integration lane pins the bytes with a literal
+// of its own, which is the point of that copy.
 //
 // Two callers need the same string for opposite reasons: this file WRITES it
 // onto every record (the natural key and activity.source), and the transport
-// directory PUBLISHES it with a label so a reader of the capture trace sees a
-// name instead of `ext:dispact-connector:dispact`. A second spelling of the
-// grammar would resolve labels for ids no record ever carries while the ones
-// that do arrive stay raw — a mismatch nothing fails on, because a directory
-// miss is a fallback rather than an error.
+// directory PUBLISHES it with a label so a reader of the capture trace has a
+// name to see. A second spelling of the grammar would resolve labels for ids no
+// record ever carries while the ones that do arrive stay raw — a mismatch
+// nothing fails on, because a directory miss is a fallback rather than an error.
 func extSourceSystem(unit, system string) string {
 	return "ext:" + unit + ":" + system
 }
